Change AllSamples to AllSamplesPage
