+++ Samples/WinMLSamplesGallery/WinMLSamplesGallery/MainWindow.xaml.cs
@@ -21,7 +21,7 @@ namespace WinMLSamplesGallery
     {
         private readonly List<(string Tag, Type Page)> _pages = new List<(string Tag, Type Page)>
         {
-            ("all_samples", typeof(AllSamples)),
+            ("all_samples", typeof(AllSamplesPage)),
             ("audio", typeof(Audio)),
             ("benchmark", typeof(Benchmark)),
             ("home", typeof(HomePage)),
